# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "CountMax Common API",
        "contact": {
            "name": "CountMax",
            "url": "https://git.countmax.ru/countmax/commonapi",
            "email": "1020@watcom.ru"
        },
        "license": {},
        "version": "2.0"
    },
    "paths": {
...
        "/v2/entities": {
            "get": {
                "description": "get slice of entities with offset, limit parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reference"
                ],
                "summary": "Get all entities types",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "default=0",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default=10",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/infra.ReferenceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    }
                }
            }
        },
        "/v2/entities/{id}": {
            "get": {
                "description": "get entity by entity ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reference"
                ],
                "summary": "Get entity by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "entity id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Entity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    }
                }
            }
        },
...
        "/v2/tasks/controllers/{sn}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get slice of tasks (json) for controller by serial number with limit parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "intraservice"
                ],
                "summary": "Get all tasks for controller's serial number",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "default=10",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SerialNumber of controller",
                        "name": "sn",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/infra.SDTasksResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    }
                }
            }
        },
        "/v2/tasks/{id}/comment": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add comment to the Task in the ServiceDesk by TaskID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "intraservice"
                ],
                "summary": "Add comment to task",
                "parameters": [
                    {
                        "description": "comment content",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.TaskComment"
                        }
                    },
                    {
                        "type": "string",
                        "description": "TaskID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/infra.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    }
                }
            }
        },
        "/v2/tasks/{id}/status": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Change status of the Task in the ServiceDesk by TaskID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "intraservice"
                ],
                "summary": "Change status task",
                "parameters": [
                    {
                        "description": "status data description",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.TaskStatus"
                        }
                    },
                    {
                        "type": "string",
                        "description": "TaskID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/infra.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/infra.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/infra.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.ControllerTask": {
            "type": "object",
            "properties": {
                "task": {
                    "type": "string"
                }
            }
        },
        "domain.ControllerTasks": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/domain.ControllerTask"
            }
        },
...
        "domain.Entities": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/domain.Entity"
            }
        },
        "domain.Entity": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "domain.TaskComment": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                }
            }
        },
        "domain.TaskStatus": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "isPrivateComment": {
                    "type": "boolean"
                },
                "resultFieldName": {
                    "type": "string"
                },
                "resultFieldValue": {
                    "type": "string"
                },
                "statusId": {
                    "type": "integer"
                }
            }
        },
...
        "infra.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "application-specific error code",
                    "type": "integer"
                },
                "error": {
                    "description": "application-level error message, for debugging",
                    "type": "string"
                },
                "status": {
                    "description": "user-level status message",
                    "type": "string"
                }
            }
        },
        "infra.HTTPError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "object"
                }
            }
        },
...
        "infra.ReferenceResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/domain.Entities"
                },
                "result_set": {
                    "type": "object",
                    "$ref": "#/definitions/infra.ResultSet"
                }
            }
        },
        "infra.ResultSet": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "infra.SDTasksResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/domain.ControllerTasks"
                },
                "result_set": {
                    "type": "object",
                    "$ref": "#/definitions/infra.ResultSet"
                }
            }
        },
        "infra.SuccessResponse": {
            "type": "object",
            "properties": {
                "httpcode": {
                    "description": "http response status code",
                    "type": "integer"
                },
                "message": {
                    "description": "text of message",
                    "type": "string"
                },
                "status": {
                    "description": "user-level status message",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}